Run tests on Github Actions
