+++ .github/workflows/build-rust.yml
@@ -17,21 +17,24 @@ jobs:
           platform: Linux
           packages:
           - gcc-aarch64-linux-gnu
-          targets:
+          test-targets:
           - x86_64-unknown-linux-gnu
+          targets:
           - aarch64-unknown-linux-gnu
         - os: macos-12
           platform: macOS
-          targets:
+          test-targets:
           - x86_64-apple-darwin
+          targets:
           - aarch64-apple-darwin
           - aarch64-apple-ios
           - aarch64-apple-ios-sim
           - x86_64-apple-ios
         - os: windows-2022
           platform: Windows
-          targets:
+          test-targets:
           - x86_64-pc-windows-msvc
+          targets:
           - aarch64-pc-windows-msvc
     runs-on: ${{ matrix.os }}
     env:
@@ -59,7 +62,7 @@ jobs:
         targets: ${{ join(matrix.targets, ', ') }}
     - name: Build
       shell: bash
-      run: cargo build --verbose --workspace --all-features --target ${{ join(matrix.targets, ' --target ') }}
-    - name: Post-Build Tests
+      run: cargo build --verbose --workspace --all-features --target ${{ join(matrix.targets, ' --target ') }} --target ${{ join(matrix.test-targets, ' --target ') }}
+    - name: Test
       shell: bash
-      run: cargo test
+      run: cargo test --verbose --workspace --all-features --target ${{ join(matrix.test-targets, ' --target ') }}
